cleanup the server units as a cherry on top
