+++ fibre/server_upload_test.go
@@ -1,9 +1,7 @@
 package fibre_test
 
 import (
-	"context"
 	"crypto/ed25519"
-	"fmt"
 	"testing"
 	"time"
 
@@ -12,13 +10,10 @@ import (
 	"github.com/celestiaorg/celestia-app-fibre/v6/x/fibre/types"
 	"github.com/celestiaorg/rsema1d"
 	"github.com/celestiaorg/rsema1d/field"
-	"github.com/cometbft/cometbft/crypto"
-	cmtproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	core "github.com/cometbft/cometbft/types"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	txsigning "github.com/cosmos/cosmos-sdk/types/tx/signing"
 	"github.com/stretchr/testify/require"
-	"google.golang.org/grpc"
 )
 
 // TestServerUploadShard unit tests the [Server.UploadShard].
@@ -139,42 +134,6 @@ func TestServerUploadShard(t *testing.T) {
 	}
 }
 
-// makeTestServer creates a server with all necessary test infrastructure.
-func makeTestServer(t *testing.T) (*fibre.Server, validator.Set, *core.Validator) {
-	t.Helper()
-
-	// create validator set (use enough validators for good distribution)
-	validators, privKeys := makeTestValidators(t, 100)
-	valSet := validator.Set{
-		ValidatorSet: core.NewValidatorSet(validators),
-		Height:       100,
-	}
-
-	// use first validator as the server's identity
-	privVal := newTestPrivValidator(privKeys[0])
-
-	// find the server validator in the ValidatorSet by matching the address
-	// Note: core.NewValidatorSet may reorder validators, so we can't assume validators[0] == privKeys[0]
-	serverPubKey, err := privVal.GetPubKey()
-	require.NoError(t, err)
-	serverAddress := serverPubKey.Address()
-
-	serverValidator, found := valSet.GetByAddress(serverAddress)
-	require.True(t, found, "server validator not found in validator set")
-	require.NotNil(t, serverValidator, "server validator is nil")
-
-	// create server
-	server, err := fibre.NewInMemoryServer(
-		privVal,
-		&mockQueryClient{},
-		&mockValidatorSetGetter{set: valSet},
-		fibre.DefaultServerConfig(),
-	)
-	require.NoError(t, err)
-
-	return server, valSet, serverValidator
-}
-
 // makeTestRequest creates a valid UploadShardRequest for the given test setup.
 // Optional modifier can be provided to customize the request after construction.
 // The promise is automatically re-signed after modification.
@@ -268,72 +227,4 @@ func makeTestRequest(
 	}
 
 	return req
-}
-
-// mockQueryClient is a mock implementation of types.QueryClient for testing.
-type mockQueryClient struct{}
-
-func (m *mockQueryClient) Params(ctx context.Context, in *types.QueryParamsRequest, opts ...grpc.CallOption) (*types.QueryParamsResponse, error) {
-	return &types.QueryParamsResponse{}, nil
-}
-
-func (m *mockQueryClient) EscrowAccount(ctx context.Context, in *types.QueryEscrowAccountRequest, opts ...grpc.CallOption) (*types.QueryEscrowAccountResponse, error) {
-	return &types.QueryEscrowAccountResponse{}, nil
-}
-
-func (m *mockQueryClient) Withdrawals(ctx context.Context, in *types.QueryWithdrawalsRequest, opts ...grpc.CallOption) (*types.QueryWithdrawalsResponse, error) {
-	return &types.QueryWithdrawalsResponse{}, nil
-}
-
-func (m *mockQueryClient) IsPaymentProcessed(ctx context.Context, in *types.QueryIsPaymentProcessedRequest, opts ...grpc.CallOption) (*types.QueryIsPaymentProcessedResponse, error) {
-	return &types.QueryIsPaymentProcessedResponse{}, nil
-}
-
-func (m *mockQueryClient) ValidatePaymentPromise(ctx context.Context, in *types.QueryValidatePaymentPromiseRequest, opts ...grpc.CallOption) (*types.QueryValidatePaymentPromiseResponse, error) {
-	// Calculate expiration time: creation_timestamp + 1 hour (default timeout)
-	expirationTime := in.Promise.CreationTimestamp.Add(1 * time.Hour)
-	currentTime := time.Now()
-
-	// Check if payment promise has expired
-	if currentTime.After(expirationTime) || currentTime.Equal(expirationTime) {
-		return nil, fmt.Errorf("payment promise expired: creation_timestamp %v + timeout %v = %v, current_time: %v", in.Promise.CreationTimestamp, 1*time.Hour, expirationTime, currentTime)
-	}
-
-	return &types.QueryValidatePaymentPromiseResponse{
-		IsValid:        true,
-		ExpirationTime: &expirationTime,
-	}, nil
-}
-
-// testPrivValidator is a simple mock PrivValidator for testing.
-type testPrivValidator struct {
-	privKey crypto.PrivKey
-}
-
-func newTestPrivValidator(privKey crypto.PrivKey) *testPrivValidator {
-	return &testPrivValidator{privKey: privKey}
-}
-
-func (m *testPrivValidator) GetPubKey() (crypto.PubKey, error) {
-	return m.privKey.PubKey(), nil
-}
-
-func (m *testPrivValidator) SignRawBytes(chainID, uniqueID string, rawBytes []byte) ([]byte, error) {
-	signBytes, err := core.RawBytesMessageSignBytes(chainID, uniqueID, rawBytes)
-	if err != nil {
-		return nil, err
-	}
-	return m.privKey.Sign(signBytes)
-}
-
-func (m *testPrivValidator) SignVote(chainID string, vote *cmtproto.Vote) error {
-	return nil
-}
-
-func (m *testPrivValidator) SignProposal(chainID string, proposal *cmtproto.Proposal) error {
-	return nil
-}
-
-func (m *testPrivValidator) GetAddress() core.Address {
-	return m.privKey.PubKey().Address()
 }
